Refactor Surrogates
