+++ draft/Random_GP.py
@@ -1,8 +1,7 @@
 import numpy as np
 import matplotlib.pyplot as plt
-from profit.sur.backend.gp_functions_old import invert, nll, predict_f, \
-    get_marginal_variance_BBQ, wld_get_marginal_variance
-from profit.sur.backend.kernels import kern_sqexp
+from profit.sur.gp.backend import invert, predict_f
+from profit.sur.gp.backend import kern_sqexp
 from profit.util.halton import halton
 
 def f(x): return x*np.cos(10*x)
