+++ draft/test_sur.py
@@ -3,7 +3,7 @@ Created: Mon Jul  8 12:12:59 2019
 @author: Christopher Albert <albert@alumni.tugraz.at>
 """
 
-from profit.sur.gaussian_process import GPSurrogate
+from profit.sur.gp.gaussian_process import GPSurrogate
 
 a = GPSurrogate()
