+++ profit/sur/__init__.py
@@ -1,4 +1,1 @@
-from . import sur
-from . import gaussian_process
-
 from .sur import Surrogate
