+++ profit/sur/gp/__init__.py
@@ -1,0 +1,4 @@
+from .gaussian_process import GaussianProcess
+from .gpy_surrogate import GPySurrogate
+from .custom_surrogate import GPSurrogate
+from .sklearn_surrogate import SklearnGPSurrogate
