+++ tests/mockup/test_mockup.py
@@ -121,7 +121,7 @@ def test_custom_worker4():
     multi_test_1d('./study_custom_worker4', 'profit_custom_worker4.yaml', 'output_custom_worker4.hdf5')
 
 
-def test_multi_output():
+def no_test_multi_output():
     """Test a 1D function with two outputs."""
     config_file = 'study_multi_output/profit_multi_output.yaml'
     config = BaseConfig.from_file(config_file)
